Move DeepEquals and References from extensions to Expression class
