+++ package.json
@@ -1,51 +1,95 @@
 {
-  "private": true,
-  "name": "botbuilder-packages",
-  "scripts": {
-    "postinstall": "lerna bootstrap --hoist",
-    "build": "lerna run build",
-    "clean": "lerna run clean",
-    "functional-test": "lerna run build && nyc mocha \"libraries/functional-tests/tests/*.test.js\"",
-    "browser-functional-test": "cd libraries/browser-functional-tests && node nightwatch.js -e",
-    "test": "lerna run build && nyc mocha \"libraries/@(adaptive*|bot*)/tests/**/*.test.js\"",
-    "test:consumer": "node ./run-consumer-test.js",
-    "test:coveralls": "lerna run build && nyc mocha \"libraries/@(adaptive*|bot*)/tests/**/*.test.js\" && nyc report --reporter=text-lcov | coveralls",
-    "test-coverage": "nyc mocha \"libraries/@(adaptive*|bot*)/tests/**/*.test.js\"",
-    "upload-coverage": "nyc report --reporter=text-lcov | coveralls",
-    "build-docs": "lerna run build-docs",
-    "eslint": "eslint  ./libraries/*/src/*.ts ./libraries/*/src/**/*.ts",
-    "eslint-fix": "eslint  ./libraries/*/src/*.ts ./libraries/*/src/**/*.ts --fix",
-    "set-dependency-versions": "node tools/util/updateDependenciesInPackageJsons.js ./libraries ${Version} ${PreviewPackageVersion} adaptive-expressions botbuilder-lg botframework-streaming botbuilder botbuilder-ai botbuilder-dialogs botbuilder-core botbuilder-applicationinsights botbuilder-testing botframework-connector botframework-config botframework-schema testbot && node tools/util/updateDependenciesInPackageJsons.js ./transcripts ${Version} botbuilder botbuilder-ai botbuilder-dialogs",
-    "update-versions": "lerna run set-version && npm run set-dependency-versions"
-  },
-  "dependencies": {
-    "@azure/ms-rest-js": "^1.8.13",
-    "@types/jsonwebtoken": "^8.3.2",
-    "@types/lodash": "^4.14.134",
-    "@typescript-eslint/eslint-plugin": "^1.10.2",
-    "@typescript-eslint/parser": "^1.10.2",
-    "coveralls": "^3.0.4",
-    "eslint": "^5.16.0",
-    "eslint-plugin-only-warn": "^1.0.1",
-    "lerna": "^3.2.1",
-    "mocha": "^6.1.4",
-    "ms-rest-azure": "^2.6.0",
-    "read-text-file": "^1.1.0",
-    "replace-in-file": "^4.1.0",
-    "sinon": "^7.3.2",
-    "typedoc": "^0.15.0",
-    "typedoc-plugin-external-module-name": "^2.1.0",
-    "typedoc-plugin-markdown": "^2.2.10"
-  },
-  "nyc": {
-    "exclude": [
-      "**/botframework*/**/generated/**",
-      "**/botbuilder*/**/generated/**",
-      "**/adaptive-expressions/**/generated/**",
-      "**/botframework-luis/**",
-      "**/tests/**",
-      "**/tools/**",
-      "**/internal.*"
-    ]
-  }
+	"private": true,
+	"name": "botbuilder-packages",
+	"dependencies": {
+		"@azure/cognitiveservices-luis-runtime": "2.0.0",
+		"@azure/cosmos": "^3.3.1",
+		"@azure/ms-rest-js": "^1.8.13",
+		"@microsoft/recognizers-text-choice": "1.1.4",
+		"@microsoft/recognizers-text-data-types-timex-expression": "^1.1.4",
+		"@microsoft/recognizers-text-date-time": "1.1.4",
+		"@microsoft/recognizers-text-number": "1.1.4",
+		"@microsoft/recognizers-text-suite": "1.1.4",
+		"@types/chai": "^4.1.7",
+		"@types/documentdb": "^1.10.5",
+		"@types/jsonwebtoken": "^8.3.2",
+		"@types/jspath": "^0.4.0",
+		"@types/lodash": "^4.14.134",
+		"@types/lru-cache": "^5.1.0",
+		"@types/mocha": "^2.2.47",
+		"@types/moment-timezone": "^0.5.12",
+		"@types/node": "^10.12.18",
+		"@types/node-fetch": "^2.3.2",
+		"@types/restify": "^7.2.1",
+		"@types/semaphore": "^1.1.0",
+		"@types/uuid": "^3.4.3",
+		"@types/ws": "^6.0.3",
+		"@types/xmldom": "^0.1.29",
+		"@typescript-eslint/eslint-plugin": "^1.10.2",
+		"@typescript-eslint/parser": "^1.10.2",
+		"adal-node": "0.2.1",
+		"antlr4ts": "0.5.0-alpha.1",
+		"applicationinsights": "1.2.0",
+		"assert": "^1.4.1",
+		"axios": "^0.19.0",
+		"azure-storage": "2.10.2",
+		"base64url": "^3.0.0",
+		"chai": "^4.2.0",
+		"chatdown": "^1.0.2",
+		"cls-hooked": "^4.2.2",
+		"codelyzer": "^4.1.0",
+		"coveralls": "^3.0.4",
+		"documentdb": "1.14.5",
+		"dotenv": "^6.1.0",
+		"eslint": "^5.16.0",
+		"eslint-plugin-only-warn": "^1.0.1",
+		"filenamify": "^4.1.0",
+		"flat": "^4.0.0",
+		"form-data": "^2.3.3",
+		"fs-extra": "^7.0.1",
+		"globalize": "^1.4.2",
+		"jsonwebtoken": "8.0.1",
+		"jspath": "^0.4.0",
+		"lerna": "^3.2.1",
+		"lodash": "^4.17.11",
+		"lru-cache": "^5.1.1",
+		"mocha": "^6.1.4",
+		"mocha-logger": "^1.0.6",
+		"moment": "^2.20.1",
+		"moment-timezone": "^0.5.25",
+		"ms-rest-azure": "^2.6.0",
+		"nock": "^10.0.3",
+		"node-fetch": "^2.2.1",
+		"nyc": "^15.0.0",
+		"path": "^0.12.7",
+		"read-text-file": "^1.1.0",
+		"replace-in-file": "^4.1.0",
+		"restify": "^8.3.0",
+		"rsa-pem-from-mod-exp": "^0.8.4",
+		"semaphore": "^1.1.0",
+		"should": "^13.2.3",
+		"sinon": "^7.3.2",
+		"source-map-support": "^0.5.3",
+		"swagger-client": "^2.1.18",
+		"ts-node": "^4.1.0",
+		"typedoc": "^0.15.0",
+		"typedoc-plugin-external-module-name": "^2.1.0",
+		"typedoc-plugin-markdown": "^2.2.10",
+		"typescript": "3.5.3",
+		"unzipper": "^0.10.9",
+		"url-parse": "^1.4.4",
+		"uuid": "^3.3.2",
+		"ws": "^7.1.2"
+	},
+	"nyc": {
+		"exclude": [
+			"**/botframework*/**/generated/**",
+			"**/botbuilder*/**/generated/**",
+			"**/adaptive-expressions/**/generated/**",
+			"**/botframework-luis/**",
+			"**/tests/**",
+			"**/tools/**",
+			"**/internal.*"
+		]
+	}
 }
